Restructure to match project details
